Minimal Dependency Mapping Implementation
This implements dependency mapping for publishing POMs, meaning that for any given declared dependency,
we publish a mapped version of that dependency. We resolve all dependencies for a variant using a secondary
resolvable configuration that has the same dependencies as the consumable configuration being published.

We can cross-reference the declared dependencies with the first-level edges of a resolution result. This
lets us know what variant each dependency resolves to. Then, we publish the resolved coordinates.

This specifically does not handle the case where non-consumable configurations are published, as is done
by AGP and KGP. This would require an extra round of attribute matching, which while possible, is not
implemented in this PR. The extra round of matchin is by nature inefficient, since we already perform matching
on consumable configurations. Furthermore such an implementation would require significant code duplication.

An internal API is added to leverage this new functionality. When adding a variant to a component,
an additional resolution configuration can be specified. Once this new functionality is properly tested
and stabilized, we will make the API public.

This new implementation is able to perform "standard" version mapping, as it exists currently. We can
also switch a flag in the future to implement the existing version mapping API with the new functionality.

This means that this commit aims to have no behavior changes. Dependency mapping is hidden behind an
internal API and is disabled by default. Furthermore, the existing version mapping code-path remains intact.